folders for ease, and moving verify-references out of the server actions and placing it as util as this makes more sense
